Fix status switching between FINISHED and DOWNLOADING
Do not change the status from an update, if it is in the status
DOWNLOADING. Reason: The FINISHED status is overwritten with
DOWNLOADING, if the recording is downloading. But during the update of
the recordings, it gets reset with the FINISHED status from the server
until it is overriden again.