Add login dialog for Cam4
The login dialog uses a WebView to load the login form from cam4.com
Username and password get filled in by ctbrec, the user has to solve
the captcha. Afterwards all cookies (especially the session cookie)
are extracted from the webview and transfered to the OkHttp client,
so that it can be used for further requests.