Add column which opens a preview popup, when hovered over
Add a column to the recorded models table, which can be used to
open a small preview popup. The popup opens, when the mouse hovers over
the table cell for a certain amount of time or if the cell gets clicked.
The preview plays the stream with the lowest quality without audio.