Add event processing system to the server
Similar to the events and actions in the client you can add event
listeners on the server. Easiest way to create them is to run ctbrec in
standalone mode and then create the event on the settings tab.
Afterwards you can copy the event handler from the client settings file
to the server settings.